requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.2.1
django-phonenumber-field==7.1.0
filelock==3.12.0
identify==2.5.24
idna==3.4
nodeenv==1.8.0
phonenumbers==8.13.11
platformdirs==3.5.1
pre-commit==3.3.1
python-decouple==3.8
PyYAML==6.0
requests==2.30.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.0
